RAG_Application_Haystack_MistralAI_Pinecone/requirements.txt
pinecone-haystack
haystack-ai
fastapi
uvicorn
python-dotenv
pathlib
setuptools
pypdf
sentence-transformers
python-multipart

-e .